lpr-microservice/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
paddlepaddle
paddleocr==2.7.3
opencv-python-headless>=4.8.0
numpy<2.0.0
Pillow>=10.0.0
PyWavelets>=1.4.0
pydantic>=2.5.0
requests>=2.31.0
